feat: dynamic OG meta tags for shared comparison links (fixes #10)
When someone shares a CompareCart comparison on WhatsApp/X/Facebook, the
social preview now shows:
- Title: "Sony WH-1000XM5 vs Bose 700 — Product Comparison | CompareCart"
- Description: "Side-by-side comparison of 3 products: [titles]..."

Implementation:
- Split compare page into server component (page.tsx) + client (CompareClient.tsx)
- page.tsx exports generateMetadata that decodes the payload server-side
- Extracted shared decodePayload utility to src/lib/decode-payload.ts
- OG + Twitter card meta tags set dynamically per comparison
